Rename `tokens.py` to `token_type`.py
